+++ internal/conn/config.go
@@ -11,6 +11,5 @@ import (
 type Config interface {
 	DialTimeout() time.Duration
 	Trace() *trace.Driver
-	ConnectionTTL() time.Duration
 	GrpcDialOptions() []grpc.DialOption
 }