+++ tests/integration/discovery_test.go
@@ -46,8 +46,7 @@ func TestDiscovery(t *testing.T) {
 				t.Fatalf("unknown request type: %s", requestTypes[0])
 			}
 		}
-		parking = make(chan struct{})
-		ctx     = xtest.Context(t)
+		ctx = xtest.Context(t)
 	)
 
 	db, err := ydb.Open(ctx,
@@ -60,7 +59,6 @@ func TestDiscovery(t *testing.T) {
 			config.WithOperationCancelAfter(time.Second*2),
 		),
 		ydb.WithBalancer(balancers.SingleConn()),
-		ydb.WithConnectionTTL(time.Second*1),
 		ydb.WithMinTLSVersion(tls.VersionTLS10),
 		ydb.WithLogger(
 			newLoggerWithMinLevel(t, log.WARN),
@@ -94,13 +92,6 @@ func TestDiscovery(t *testing.T) {
 				}),
 			),
 		),
-		ydb.WithTraceDriver(trace.Driver{
-			OnConnPark: func(info trace.DriverConnParkStartInfo) func(trace.DriverConnParkDoneInfo) {
-				return func(info trace.DriverConnParkDoneInfo) {
-					parking <- struct{}{}
-				}
-			},
-		}),
 	)
 	if err != nil {
 		t.Fatal(err)
@@ -117,17 +108,5 @@ func TestDiscovery(t *testing.T) {
 		} else {
 			t.Log(endpoints)
 		}
-		t.Run("wait", func(t *testing.T) {
-			t.Run("parking", func(t *testing.T) {
-				<-parking // wait for parking conn
-				t.Run("re-discover", func(t *testing.T) {
-					if endpoints, err := db.Discovery().Discover(ctx); err != nil {
-						t.Fatal(err)
-					} else {
-						t.Log(endpoints)
-					}
-				})
-			})
-		})
 	})
 }