+++ trace/table.go
@@ -312,7 +312,6 @@ type (
 	}
 	TableInitDoneInfo struct {
 		Limit int
-		Error error
 	}
 	TablePoolStateChangeInfo struct {
 		Size  int